feat: bridge open url
