+++ src/components/Activity/ActivityClaimButton.tsx
@@ -9,8 +9,8 @@ import LoadingIcon from '@/assets/loading.svg';
 import { ActivityContext } from '@/components/Activity/ActivityContext.js';
 import { ActivityMintSuccessDialog } from '@/components/Activity/ActivityMintSuccessDialog.js';
 import { useActivityClaimCondition } from '@/components/Activity/hooks/useActivityClaimCondition.js';
-import { useIsFollowTwitterInActivity } from '@/components/Activity/hooks/useActivityFollowTwitter.js';
 import { useActivityPremiumList } from '@/components/Activity/hooks/useActivityPremiumList.js';
+import { useIsFollowTwitterInActivity } from '@/components/Activity/hooks/useIsFollowTwitterInActivity.js';
 import { classNames } from '@/helpers/classNames.js';
 import { enqueueErrorMessage } from '@/helpers/enqueueMessage.js';
 import { getSnackbarMessageFromError } from '@/helpers/getSnackbarMessageFromError.js';
